fix(eikon): sesión de 1h → 30 días + renovación deslizante (no más deslogueo)
EIKON_WEBAPP_JWT_TTL 3600→2592000 (30 días). current_user (deps.py) ahora re-emite la cookie
de sesión con exp fresco en cada request autenticado (sliding expiration) → un usuario activo
NUNCA se desloguea; solo caduca tras ~30 días de inactividad. Mismos atributos de cookie
(httpOnly/secure/samesite). +2 tests de sesión. (Incluye limpieza menor de type hints de mypy
en varios módulos, sin cambios de lógica.) 372 tests verdes.

+++ contrast_validator.py
@@ -75,7 +75,7 @@ class ContrastValidator:
         c = channel / 255.0
         if c <= 0.04045:
             return c / 12.92
-        return ((c + 0.055) / 1.055) ** 2.4
+        return float(((c + 0.055) / 1.055) ** 2.4)
 
     def _calculate_luminance(self, rgb: tuple[int, int, int]) -> float:
         """
